sprint-1/step-2 fix: перенести обрбаотку закрытия по эск в modal и удалить неиспользуемый файл
