+++ src/components/App/App.tsx
@@ -46,17 +46,6 @@ function App() {
     setIngredient({})
   }
 
-  const listenEscHandler = (event : any) => {
-    if (event.key === 'Escape') {
-      handleCloseModal()
-    }
-  }
-
-  React.useEffect(() => {
-    document.addEventListener('keydown', listenEscHandler)
-    return  () => document.removeEventListener('keydown', listenEscHandler)
-  }, [])
-
   React.useEffect(() => {
     fetch(BURGER_API_URL).then((response) => checkResponse(response))
     .then((data) => {
